pyproject.toml
[project]
name = "text2video"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
	"Pillow>=9.0.0",
	"ffmpeg-python>=0.2.0",
	"pydub>=0.25.1",
	"typing-extensions>=4.0.0",
	"openai>=0.27.0",
	"python-dotenv>=1.0.0",
	"requests>=2.28.0",
	"dashscope>=0.1.0",
]
